+++ src/preload/index.mts
@@ -1,12 +1,15 @@
 import { electronAPI } from "@electron-toolkit/preload";
 import { contextBridge } from "electron";
+import { ipcRenderer } from "electron";
 
 // Custom APIs for renderer
-const api = {};
+const api = {
+  window: {
+    getAllThumbnail: () => ipcRenderer.invoke("window::get-all-thumbnail"),
+    getStream: () => ipcRenderer.invoke("window::get-stream"),
+  },
+};
 
-// Use `contextBridge` APIs to expose Electron APIs to
-// renderer only if context isolation is enabled, otherwise
-// just add to the DOM global.
 if (process.contextIsolated) {
   try {
     contextBridge.exposeInMainWorld("electron", electronAPI);
@@ -15,8 +18,6 @@ if (process.contextIsolated) {
     console.error(error);
   }
 } else {
-  // @ts-ignore (define in dts)
   window.electron = electronAPI;
-  // @ts-ignore (define in dts)
   window.api = api;
 }
